kjkkjkl

+++ app/Http/Controllers/UsersController.php
@@ -4,6 +4,7 @@
 
     use App\Http\Requests\UserUpdateRequest;
     use App\Models\User;
+    use Illuminate\Support\Facades\Auth;
 
     class UsersController extends Controller
     {
@@ -29,6 +30,12 @@
 
         public function destroy(User $user)
         {
+            if (Auth::user()->id === $user->id) {
+                abort(403, 'You cannot delete your self');  // or return back()->with('error', 'You cannot delete yourself');  // or return redirect()->route('users.index')->with('error', 'You cannot delete yourself');  // or return redirect()->back()->with('error', 'You cannot delete yourself');  // or return redirect()->route('dashboard')->with('error', 'You cannot delete yourself');  // or return redirect()->route('dashboard', ['error' => 'You cannot delete yourself']);  // or return redirect()->route('dashboard', ['message' => 'You cannot delete yourself', 'type' => 'error']);  // or return redirect()->route('dashboard')->withErrors(['error' => 'You cannot delete yourself']);  // or return redirect()->route('dashboard')->withInput()->withErrors(['error' => 'You cannot delete yourself']);  // or return redirect()->route('dashboard')->withInput()->with(['error' => '    You cannot delete yourself']); // or return redirect()->route('dashboard
+                //
+                //  ')
+            }
+
             $user->delete();
 
             return redirect()->route('users.index');
